Move unit test to the same package as its source
